separate vde migration to own package
